13 April 2012, 04:55 AM
Dropped LOTS of obsolete code right after this tag, including getting rid of Experiment (finally!) and lots and lots of other obsolete, moribund, or otherwise unnecessary code.

Also started on integrating the new layers code which uses data binding and hierarchical data templates, and also started to bring in the new database-driven way of managing our scenario and environment data.

Big, big changes.  If you're looking for any of that old code, this is the place to look.